requirements.txt
matplotlib
numpy
pandas
python-telegram-bot>=20
seaborn
